Changing the way the database handles warehouse items, and changing /scripts/addcart.php accordingly. Preventing the deletion of the currently logged in user. Adding /pages/warehouse.php to manage warehouse items and their quantities, and placeholder /scripts/editwh.php. Updating TODO.

+++ scripts/deleteuser.php
@@ -27,6 +27,12 @@
     exit();
   }
 
+  if($_SESSION["loggedIn"]["user_ID"] == $_GET["userid"])
+  {
+    $_SESSION["error"] = "Nie możesz usunąć swojego konta!";
+    exit();
+  }
+
   $sql = "DELETE FROM users WHERE id = $_GET[userid]";
   $conn->query($sql);
   if ($conn->affected_rows == 1)
